fix mailbox (delperm)

+++ Commands/Public/mailbox.js
@@ -9,6 +9,8 @@ module.exports = async(client, msg, suffix) => {
 	let perm = msg.guild ? msg.guild.members.cache.get(msg.author.id).hasPermission("MANAGE_GUILD") : true;
 	if (!perm) perm = msg.author.support;
 
+	const delperm = msg.guild ? msg.channel.permissionsFor(client.user).has("MANAGE_MESSAGES") : false
+
 	// get their mailbox
 	let mailbox = await r.table("Mailbox").get(msg.channel.id);
 	let omsg,
@@ -37,7 +39,7 @@ module.exports = async(client, msg, suffix) => {
 		}
 
 		omsg.delete().catch(e => null);
-		collected.delete().catch(e => null);
+		if (delperm) collected.delete().catch(e => null);
 		if (/^no$/i.test(collected.content)) {
 			msg.author.busy = false;
 			return;
@@ -62,7 +64,7 @@ module.exports = async(client, msg, suffix) => {
 			return msg.channel.send({ embed: { color: config.colors.error, title: "Timed out", description: "You ran out of time, get an autoreply ready and start the set-up again." } });
 		}
 
-		collected.delete().catch(e => null);
+		if (delperm) collected.delete().catch(e => null);
 
 		// Succesful autoreply
 		let autoreply = collected.content;
@@ -124,7 +126,7 @@ module.exports = async(client, msg, suffix) => {
 		msg.author.busy = false;
 		if (!collected) return;
 
-		collected.delete().catch(e => null);
+		if (delperm) collected.delete().catch(e => null);
 
 		// Succesful autoreply
 		let autoreply = collected.content;
@@ -178,7 +180,7 @@ module.exports = async(client, msg, suffix) => {
 				m => m.author.id === msg.author.id && (/^0$/.test(m.content) || responses.includes(m.content.toLowerCase()) || (parseInt(m.content) != page && parseInt(m.content) > 0 && parseInt(m.content) <= pages) || messages.filter(message => message.id == m.content).length > 0),
 				{	time: 120000, max: 1 })).first();
 
-			if (collected) {
+			if (collected && delperm) {
 				collected.delete().catch(e => null);
 			}	else if (!collected) {
 				msg.author.busy = false;
@@ -204,7 +206,7 @@ module.exports = async(client, msg, suffix) => {
 						{ time: 120000, max: 1 },
 					)).first();
 
-					if (collected) collected.delete().catch(e => null);
+					if (collected && delperm) collected.delete().catch(e => null);
 					if (!collected) {
 						msg.author.busy = false;
 						omsg.delete().catch(e => null);
@@ -234,7 +236,7 @@ module.exports = async(client, msg, suffix) => {
 						{ time: 120000, max: 1 },
 					)).first();
 
-					if (collected) collected.delete().catch(e => null);
+					if (collected && delperm) collected.delete().catch(e => null);
 					if (!collected) {
 						msg.author.busy = false;
 						omsg.delete().catch(e => null);
@@ -270,7 +272,7 @@ module.exports = async(client, msg, suffix) => {
 					)).first();
 
 					msg.author.busy = false;
-					if (collected) collected.delete().catch(e => null);
+					if (collected && delperm) collected.delete().catch(e => null);
 					if (!collected || /^0$/.test(collected.content)) break;
 
 					await r.table("Mailbox").get(mailbox.id).update({ autoreply: collected.content });
@@ -328,7 +330,7 @@ module.exports = async(client, msg, suffix) => {
 				return;
 			}
 
-			collected.delete().catch(e => null);
+			if (delperm) collected.delete().catch(e => null);
 
 			let index;
